+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
-## Version 1.5.0 - 2025-07-08
+## Version 1.5.0 - 2025-07-16
 
 ### Added
 
 - Adopt stable semantic conventions
+- Observation of [Persistent Queue](https://cap.cloud.sap/docs/node.js/queue#persistent-queue) metrics (beta)
+  + Disable via config `cds.requires.telemetry.metrics._queue = false`
 
 ## Version 1.4.1 - 2025-06-27